+++ intcode/InputProducer.java
@@ -2,5 +2,5 @@ package intcode;
 
 @FunctionalInterface
 public interface InputProducer {
-    int produceInput(IntCodeComputer computer);
+    long produceInput(IntCodeComputer computer);
 }
