+++ intcode/OutputListener.java
@@ -1,5 +1,5 @@
 package intcode;
 
 public interface OutputListener {
-    void outputProduced(int output);
+    void outputProduced(long output);
 }
